requirements.txt
qdrant-client
fastembed
sentence-transformers
groq
python-dotenv
transformers
torch
pypdf
pydantic